PyIface: Rename Python package to pyiface
