+++ src/cloudsc_pyiface/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "cloudscpytools"
+name = "pyiface"
 version = "0.1.0"
 authors = [
     {name = "Zbigniew Piotrowski", email = "zbigniew.piotrowski@ecmwf.int"}
