+++ src/cloudsc_pyiface/src/pyiface/__init__.py
@@ -8,5 +8,5 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
-from cloudscpytools.cloudsc_data import *  # noqa
-from cloudscpytools.dynload import *  # noqa
+from pyiface.cloudsc_data import *  # noqa
+from pyiface.dynload import *  # noqa
